日记列表受隐私管理

+++ plugins/notes/list.php
@@ -165,21 +165,21 @@ if (isset($user) && isset($avtor['id'])) $frend = dbresult(dbquery("SELECT COUNT
 title();
 aut(); // 授权表格
 err();
-if ($notes['private'] == 1 && $user['id'] != $notes['id_user'] && $frend != 2  && !user_access('notes_delete')) {
+if (empty($user) && $notes['private'] == 1 && $user['id'] != $notes['id_user'] && $frend != 2  && !user_access('notes_delete')) {
 	msg('日记只提供给朋友');
 	echo "  <div class='foot'>";
 	echo "<a href='index.php'>返回</a><br />";
 	echo "   </div>";
 	include_once '../../sys/inc/tfoot.php';
 }
-if ($notes['private'] == 2 && $user['id'] != $notes['id_user']  && !user_access('notes_delete')) {
+if (empty($user) && $notes['private'] == 2 && $user['id'] != $notes['id_user']  && !user_access('notes_delete')) {
 	msg('用户已禁止查看日记');
 	echo "  <div class='foot'>";
 	echo "<a href='index.php'>返回</a><br />";
 	echo "   </div>";
 	include_once '../../sys/inc/tfoot.php';
 }
-if (isset($_GET['delete']) && ($user['id'] == $notes['id_user'] || user_access('notes_delete'))) {
+if (isset($user) && isset($_GET['delete']) && ($user['id'] == $notes['id_user'] || user_access('notes_delete'))) {
 	echo "<center>";
 	echo "你真的想删除日记吗 " . output_text($notes['name']) . "?<br />";
 	echo "[<a href='delete.php?id=$notes[id]'><img src='/style/icons/ok.gif'> 删除</a>] [<a href='list.php?id=$notes[id]'><img src='/style/icons/delete.gif'> 取消</a>] ";
